Refactor and reorganize entire repo, no longer immitating YGS2k
From now on, we're doing our own thing, no longer making any effort to
have our code look like a YGS2k codebase. And, we're going to be SDL-only,
so SDL APIs are fair game to use anywhere in the codebase.

This is only the start of breaking away from the YGS2k legacy. For example,
some of the APP_* functions might go away later, like some of the audio functions,
and some things might be changed, renamed, or even removed.