Tomoyo TI EX7 stage record was not saved
EX7 stage record was not saved because it was overwritten by recFaultTime(). As far as I can tell, this function doesn't actually serve any purpose. The stage times are initialized as zero and times of zero are ignored by the ranking code in sectime.c, and stage_time isn't used anywhere else.